pyproject.toml
[project]
name = "langchain-course"
version = "0.1.0"
description = "Demo ongoing undertaking using langchain-langgraph"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "black>=26.3.1",
    "isort>=8.0.1",
    "langchain>=1.2.15",
    "langchain-core>=1.2.28",
    "langchain-ollama>=1.1.0",
    "langchain-openai>=1.1.12",
    "langchain-tavily>=0.2.18",
    "langgraph>=1.1.6",
    "pydantic>=2.13.0",
    "python-dotenv>=1.2.2",
    "tavily-python>=0.7.23",
]
